using proper libc primitive types for inputs and outputs

+++ src/embed.rs
@@ -1,16 +1,16 @@
 extern crate libc;
 
-use libc::c_char;
+use libc::{c_char,c_int, c_float};
 use std::ffi::{CStr,CString};
 use std::vec::{Vec};
 
 #[no_mangle]
-pub extern fn rs_int_in_int_out(input: i32) -> i32{
+pub extern fn rs_int_in_int_out(input: c_int) -> c_int{
     input*2
 }
 
 #[no_mangle]
-pub extern fn rs_string_in_string_out(s_raw: *const c_char, out: *mut c_char) -> usize {
+pub extern fn rs_string_in_string_out(s_raw: *const c_char, out: *mut c_char) -> c_int {
     // take string from the input C string
     if s_raw.is_null() { return 0; }
 
@@ -30,42 +30,42 @@ pub extern fn rs_string_in_string_out(s_raw: *const c_char, out: *mut c_char) ->
     unsafe{ std::ptr::copy(c_result.unwrap().as_ptr(), out, len); };
 
     // return result length
-    return len;
+    return len as c_int;
 }
 
 #[no_mangle]
-pub extern fn rs_numeric_array_in_numeric_array_out(src_raw: *const u32, dst: *mut u32, size: usize) -> usize {
+pub extern fn rs_numeric_array_in_numeric_array_out(src_raw: *const c_int, dst: *mut c_int, size: c_int) -> c_int {
     // Convert to borrowed pointers.
-    let src = unsafe { std::slice::from_raw_parts(src_raw, size) };
-    let mut res: Vec<u32> = Vec::with_capacity(10);
+    let src = unsafe { std::slice::from_raw_parts(src_raw, size as usize) };
+    let mut res: Vec<c_int> = Vec::with_capacity(10);
 
     res.push(42);
-    for i in 0..size {
+    for i in 0..(size as usize) {
         res.push(src[i]*2);
     }
     let res_size = res.len();
 
     unsafe{ std::ptr::copy(res.as_ptr(), dst, res_size); };
 
-    return res_size;
+    return res_size as c_int;
 }
 
 
 #[repr(C)]
 pub struct SomeStruct {
-    some_item: i32,
-    another_item: i32,
-    test: i32,
-    float_item: f32
+    some_item: c_int,
+    another_item: c_int,
+    test: c_int,
+    float_item: c_float
 }
 
 #[no_mangle]
 pub extern fn rs_struct_out() -> SomeStruct {
     let sss = SomeStruct {
-        some_item: 3,
-        another_item: 4,
-        test: 5,
-        float_item: 0.7_f32
+        some_item: 3 as c_int,
+        another_item: 4 as c_int,
+        test: 5 as c_int,
+        float_item: 0.7_f32 as c_float
     };
     return sss;
 }
@@ -73,8 +73,8 @@ pub extern fn rs_struct_out() -> SomeStruct {
 
 #[repr(C)]
 pub struct OtherStruct {
-  int_setting: i32,
-  float_setting: f32,
+  int_setting: c_int,
+  float_setting: c_float,
   bool_setting: bool
 }
